El user ahora puede agregar camaras clickeando sobre el mapa :D

+++ rustic_city_eye/src/surveilling/location.rs
@@ -1,14 +1,14 @@
 #[derive(Debug)]
 #[allow(dead_code)]
 pub struct Location {
-    longitude: u16,
-    latitude: u16,
+    longitude: f64,
+    latitude: f64,
 }
 
 impl Location {
     pub fn new(lat: String, long: String) -> Location {
-        let latitude = lat.parse::<u16>().unwrap();
-        let longitude = long.parse::<u16>().unwrap();
+        let latitude = lat.parse::<f64>().unwrap();
+        let longitude = long.parse::<f64>().unwrap();
 
         Location {
             longitude,
